% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghoneycomb-package.R
\docType{package}
\name{gghoneycomb-package}
\alias{gghoneycomb}
\alias{gghoneycomb-package}
\title{gghoneycomb: Honeycomb Plots for ggplot2}
\description{
Provides geoms and stats for creating honeycomb-style hexagonal grid visualizations as an extension to ggplot2. Enables the display of categorical or continuous data in a compact, visually appealing hexagonal layout.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dleopold/gghoneycomb}
  \item \url{https://dleopold.github.io/gghoneycomb/}
  \item Report bugs at \url{https://github.com/dleopold/gghoneycomb/issues}
}

}
\author{
\strong{Maintainer}: Devin R Leopold \email{devin.leopold@gmail.com}

}
\keyword{internal}
