% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb_compactness.R
\name{hex-compactness}
\alias{hex-compactness}
\title{Hex region compactness utilities}
\description{
Internal functions for measuring perimeter and compactness of regions
on a hexagonal grid. Used for scoring allocator quality.
}
\keyword{internal}
