% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honey_pal.R
\name{honey_pal}
\alias{honey_pal}
\title{Honey-inspired palette}
\usage{
honey_pal(n = 6, direction = 1)
}
\arguments{
\item{n}{Number of colors to return.}

\item{direction}{Direction of the palette. Use \code{1} for dark-to-light
(default) or \code{-1} for light-to-dark.}
}
\value{
A character vector of hex color codes.
}
\description{
A warm, honey-like palette for categorical data. The palette is designed
to work well with discrete categories and can be interpolated when more
colors are needed.
}
\examples{
honey_pal()
honey_pal(3)
honey_pal(8)

}
