% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_honeycomb.R
\name{theme_honeycomb}
\alias{theme_honeycomb}
\title{Minimal theme for honeycomb plots}
\usage{
theme_honeycomb(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A clean, honeycomb-focused theme based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_void()}}. This
removes axes and grids while keeping typographic hierarchy for titles and
captions.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot() + ggplot2::theme_void() + theme_honeycomb()
}

}
