% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghoneycomb-package.R
\docType{package}
\name{gghoneycomb-package}
\alias{gghoneycomb}
\alias{gghoneycomb-package}
\title{gghoneycomb: Honeycomb Plots for ggplot2}
\description{
A ggplot2 extension for honeycomb waffle charts where each category forms a contiguous region of hexagonal cells. Supports rectangular, rounded, and organic silhouette shapes; compact and free (temperature-controlled) layout modes; 90-degree rotation; and a honey-inspired color palette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dleopold/gghoneycomb}
  \item \url{https://dleopold.github.io/gghoneycomb/}
  \item Report bugs at \url{https://github.com/dleopold/gghoneycomb/issues}
}

}
\author{
\strong{Maintainer}: Devin R Leopold \email{devin.leopold@gmail.com}

}
\keyword{internal}
